requirements.txt
dash==2.16.1
pandas==2.2.2
plotly==5.21.0
requests==2.31.0
numpy==1.26.4
gunicorn==21.2.0
python-dotenv==1.0.1